Switch to Jest's snapshot testing
fixes #35
