+++ package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "build": "tsc -p .",
     "prepublish": "npm run build",
-    "test": "mocha --compilers=ts:ts-node/register --no-timeouts test/**/test.*.ts"
+    "test": "jest --coverage"
   },
   "author": "Ryan Cavanaugh",
   "license": "Apache-2.0",
@@ -17,12 +17,24 @@
     "url": "https://github.com/RyanCavanaugh/dts-dom"
   },
   "devDependencies": {
-    "@types/chai": "^4.0.7",
-    "@types/mocha": "^2.2.44",
-    "@types/node": "^8.0.54",
-    "chai": "^4.1.2",
-    "mocha": "^4.0.1",
-    "ts-node": "^3.3.0",
+    "@types/jest": "^22.0.1",
+    "jest": "^22.1.3",
+    "ts-jest": "^22.0.1",
     "typescript": "^2.6.2"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.ts$": "ts-jest"
+    },
+    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.ts$",
+    "moduleFileExtensions": [
+      "ts",
+      "js",
+      "json"
+    ],
+    "mapCoverage": true,
+    "collectCoverageFrom": [
+      "lib/**/*.ts"
+    ]
   }
 }
